Update SoundsDownloadScript.ps1
Added the ability to store the script configuration options in an external .ps1 file which can then be used by calling the $DotSrcConfig parameter with the path to file as a value. It calls the script via dot sourcing. It will then override any of the options that are set in the script itself. There is VERY crude validation.