Updated SoundsDownloadScript.ps1 to add DebugDirectory parameter to command line options
If Debug and DebugDirectory are set either as command line parameters or in the inline configuration options, the script will turn on debug logging right away. Otherwise, it will wait to read the dot source script and run it after, if needed.

Also, moved the general functions to the top of the script.