genRSS.ps1 can now update feeds when the profile has been updated
If UseProfileHash=yes genRSS will get the hash of the profile file using Get-FileHash and put it in the XML under the <hash> element. Then every time genRSS runs, it will compare the hash in the XML to the hash of the profile file. If they don't match it will run and build a new XML file. Basically, it won't wait for a new episode to update.